Remove the include_internal include path
Instead put all private headers in the src directory, which is a more common
practice.
