+++ src/rocket_engine.h
@@ -25,7 +25,8 @@
 
 #pragma once
 
-#include <internal/rocket_future.h>
 #include <rocket/rocket_engine.h>
+
+#include "rocket_future.h"
 
 rocket_future_t* rocket_engine_await_next(rocket_engine_t* engine);
