+++ src/rocket_executor.h
@@ -25,38 +25,19 @@
 
 #pragma once
 
-#include <internal/dlist.h>
-#include <internal/pal.h>
-#include <rocket/rocket_fiber.h>
-#include <rocket/rocket_types.h>
+#include <rocket/rocket_executor.h>
 
-typedef enum {
-  NONE = 0,
-  RUNNABLE = 1,
-  BLOCKED = 2,
-  COMPLETED = 3,
-} rocket_fiber_state_t;
+#include "dlist.h"
 
-typedef struct rocket_fiber {
-  dlist_node_t list_node;
+struct rocket_executor {
+  rocket_engine_t* engine;
 
-  // State of the fiber.
-  rocket_fiber_state_t state;
-  // Executor that's current running the fiber.
-  rocket_executor_t* executor;
-  // Function running in the fiber.
-  rocket_task_func_t task_func;
-  // Context used in the function.
-  void* context;
+  // Runnable fibers.
+  dlist_node_t runnable;
+  // Blocked fibers.
+  dlist_node_t blocked;
 
-  pal_stack_t stack;
-  void* stk_ptr;
-} rocket_fiber_t;
+  void* execute_loop_stk_ptr;
+};
 
-rocket_fiber_t* rocket_fiber_create(
-    rocket_executor_t* executor,
-    rocket_task_func_t func,
-    void* context);
-rocket_fiber_t* get_current_fiber();
-void set_current_fiber(void* fiber);
-void rocket_fiber_destroy(rocket_fiber_t* fiber);
+rocket_engine_t* rocket_executor_get_engine(rocket_executor_t* executor);
